refactor: 상수 분리
[#360]
