+++ frontend/src/hooks/useServerStationFilters.ts
@@ -6,7 +6,7 @@ import {
   selectedCompanyNamesFilterStore,
 } from '@stores/selectedServerStationFiltersStore';
 
-import type { CHARGER_TYPES, CAPACITIES, COMPANY_NAME } from '@constants';
+import type { CAPACITIES, CHARGER_TYPES, COMPANY_NAME } from '@constants/chargers';
 
 export const useServerStationFilters = () => {
   const [selectedCompanyNamesFilters, setSelectedCompanyNamesFilter] = useExternalState(
